# swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "FILMESGRAM",
    "description": "Esta API tem como objetivo a criação de uma comunidade de cinéfilos por meio da interação dos usuários com os filmes cadastrados",
    "contact": {
      "email": "marcioh.dev@gmail.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/listarfilme": {
      "get": {
        "description": "Rota responsável por listar todos os filmes cadastrados no json",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/addfilme": {
      "post": {
        "description": "Rota responsável por adicionar um filme",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "example": "any"
                },
                "diretor": {
                  "example": "any"
                },
                "lancamento": {
                  "example": "any"
                },
                "genero": {
                  "example": "any"
                },
                "descricao": {
                  "example": "any"
                },
                "imagem": {
                  "example": "any"
                },
                "ondeAssistir": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/editarfilme/{id}": {
      "put": {
        "description": "Rota responsável por editar um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "example": "any"
                },
                "diretor": {
                  "example": "any"
                },
                "lancamento": {
                  "example": "any"
                },
                "genero": {
                  "example": "any"
                },
                "descricao": {
                  "example": "any"
                },
                "imagem": {
                  "example": "any"
                },
                "ondeAssistir": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/deletarfilme/{id}": {
      "delete": {
        "description": "Rota responsável por deletar um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/listarusuario": {
      "get": {
        "description": "Rota responsável por listar todos os usuários cadastrados no json",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "/adicionarusuario": {
        "post": {
          "description": "Rota responsável por adicionar um usuário",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "senha": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    },
    "/editarusuario/{id}": {
      "put": {
        "description": "Rota responsável por editar um usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "/deletarusuario{id}": {
        "delete": {
          "description": "Rota responsável por deletar um usuário",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    },
    "/usuarios/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/usuarios/verificar-token": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/listaravaliacao": {
      "get": {
        "description": "Rota responsável por listar todas as avaliações de filmes cadastradas no json",
        "parameters": [
          {
            "name": "idUsuario",
            "in": "query",
            "type": "string"
          },
          {
            "name": "idFilme",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "/addavaliacao": {
        "post": {
          "description": "Rota responsável por cadastrar uma nova avaliação de filme",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "idUsuario": {
                    "example": 3
                  },
                  "idFilme": {
                    "example": 4
                  },
                  "comentario": {
                    "example": "excelente"
                  },
                  "curtiu": {
                    "example": true
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/editaravaliacao": {
        "put": {
          "description": "Rota responsável por editar uma avaliação de filme já existente",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "idAvaliacao": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    },
    "/deletaravaliacao/{id}": {
      "delete": {
        "description": "Rota responsável por deletar uma avaliação de filme já existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}